[ADD]: Mongo yml
To run docker-compose.yml -> "docker compose up -d"

This will trigger docker to create an image with the specified version of mongodb and run the container.
Wait until you see "Waiting for connections".

Later on, with MongoDB Compass, you can connect to the DB with the variables you named on the .env