[ADD]: Prisma config
Prisma ORM, similar to mongoose for MongoDB. Will use this to operate PostgreSQL.

Steps:
- pnpm i -D prisma
- pnpm exec prisma init --datasource-provider PostgreSQL

This will create a new registry in the .env, copy it to POSTGRES_URL instead of using the prisma provided one and use the values for user, password and dbName.
Like: POSTGRES_URL="postgresql://postgres:123456@localhost:5432/NOC"

If your database already has data, you can use prisma pull to create the schema.
If not, you must create a migration with:
pnpm exec prisma migrate dev --name init

This will create the necessary configs and migrations to use prisma. This includes the DB connection, the model to use in node, etc, just like MongoDB but automatic.