[ADD]: Vitest instead of jest
Jest was kind of throwing a lot of errors because I was not using ts-node as the course did.
Opted to go for vitest, pretty much worked instantly and required not much configuration.

The docker-compose test is a new file that is just there to setup new containers that are specific for testing purposes and to not clog the og ones.

Also, needed to create a .env.test file which is ignored in the general gitignore, but in there the only thing that changes is the dbName and password for both mongo and postgre, which are the ones used in the docker-compose test.

tsconfig was edited so that I wouldn't need to import in each time the describe, test or expect functions in test files. Allowing globals, so to say.

Besides that, the vitest.config.ts is all the minimum config you need, in which it loads specifically the .env.test file instead of the normal .env

Not sure if vite is really necessary, as the docs only tells to install vitest, but later on they say it only works if vite is also there ... so, yeah, keep both. Just in case.